requirements.txt
aiogram
python-dotenv
pillow

